Added DAS and Internal Gravity speed timings
*added DAS and Internal Gravity. warning they very buggy right now
*fixed some "may lose data warnings" from the complier

+++ tetris/Block.cpp
@@ -54,7 +54,7 @@ void Block::draw(SDL_Renderer* renderer, int x, int y)
 
 Block::Block()
 {
-	BlockRect.w = blockSize / blockBorder;
-	BlockRect.h = blockSize / blockBorder;
+	BlockRect.w = (int)(blockSize / blockBorder);
+	BlockRect.h = (int)(blockSize / blockBorder);
 	BorderPxSize = blockSize - BlockRect.w;
 }
